requirements.txt
flask
gunicorn  
torch
torchvision
Pillow
huggingface_hub
numpy
groq
twilio
python-telegram-bot
transformers
google-genai